quiz_project/requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.6.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==24.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
